Use schedule name as class name if unspecified
resque-scheduler will use the name of the scheduler as the job class
name, if the name isn't specified in the 'class' attribute of the
schedule.  This was not being supported in Active Scheduler.  The
workaround is to simply specify the class name, but this change makes
Active Scheduler that much more of a drop-in.