Add ActiveScheduler.config.{guard_with,queue_name_resolver}
This allows configuration of a custom guard clause to check whether jobs should be wrapped or not, as well as allowing for a custom queue_name_resolver configuration which will be called to resolve the queue name when the queue is not configured. The idea behind this is to enable an option which avoids the use of "class_name.constantize", meaning that we don't have to load the entire Rails environment for the resque scheduler rake task and save some memory.